requirements.txt
# Ray 1.8.0 removed from PyPI; 2.31+ works with this codebase.
# [default] pulls deps needed for runtime_env (raylet agent_manager); plain `ray` can error on clusters.
ray[default]>=2.31.0,<3
# 2.0.4 rejected by pip>=24.1 (invalid requires: PyYAML>=5.1.*)
omegaconf>=2.1.0,<3
grpcio
grpcio-tools==1.48.2
protobuf==3.20.0
protobuf3-to-dict==0.1.5
pickle5~=0.0.11
numpy==2.0.2
tensorboardX~=2.1
tensorboard
readerwriterlock==1.0.9
nashpy==0.0.21
psutil~=5.8.0
# gym 0.21.0 fails to build with recent setuptools (invalid extras_require parsing).
# 0.26.x provides wheels and keeps gym.spaces API used in this repo.
gym>=0.26.0,<0.29
dataclasses==0.6
wrapt==1.12.1
pyyaml==5.4.1
dm-tree~=0.1.7
parallel-ssh==2.11.1
tabulate==0.8.10
six==1.17.0
easydict==1.9
ipython==8.18.1
colorlog==6.6.0
#hiredis==2.0.0
matplotlib==3.9.4
pandas~=1.5.1
wandb==0.25.0
requests>=2.28.0  # wandb dependency; some shared envs miss it
